+++ dbt/include/glue/macros/materializations/incremental/strategies.sql
@@ -12,8 +12,12 @@
 {% macro get_insert_into_sql(source_relation, target_relation) %}
     {%- set dest_columns = adapter.get_columns_in_relation(target_relation) -%}
     {%- set dest_cols_csv = dest_columns | map(attribute='name') | join(', ') -%}
-    insert into table {{ target_relation }}
-    select {{dest_cols_csv}} from {{ source_relation }}
+    {%- set schema_change_mode = config.get('on_schema_change', default='ignore') -%}
+    {%- if schema_change_mode != 'ignore' -%}
+    insert into table {{ target_relation }} select {{dest_cols_csv}} from {{ source_relation }}
+    {%- else -%}
+    insert into table {{ target_relation }} select {{dest_cols_csv}} from {{ source_relation.include(schema=false) }}
+    {%- endif -%}
 {% endmacro %}